exu: do not check fuType when illegalJump.valid
When illegal jump is valid, the instruction may be of any fuType.
If it's an LSU instruction, previously we have to wait for LSU output.
However, this will make the CSR redirect invisible to WBU and Frontend.
Thus, we should manully let the instruction leave in this case.

Other exceptions will not cause errors because the ITLB manully sets
the instruction to 0.U when page fault or access fault exceptions
happen.