Add support for cross-boundary access-fault
According to the RISC-V manual, the mtval will contain the virtual
address of the portion of the instruction that caused the fault,
while mepc will point to the beginning of the instruction.

Previously only cross-page page-fault is supported. This commit
renames the signal to crossBoundaryFault and supports access-fault
as well.