csr: check illegal access for SATP before SATP.MODE
We used to have the wen fix for SATP, because RISC-V requires that
satp is not modified is SATP.mode is written with an illegal value.
However, this fix bypasses the exception logic and mark the instruction
as not write-enabled. If satp is written in the U mode with an illegal
mode, the instruction may not trigger exceptions.

This commit fixes this issue.