csr: check TW bit before executing WFI instructions
When TW=1, then if WFI is executed in any less-privileged mode, and
it does not complete within an implementation-specific, bounded time
limit, the WFI instruction causes an illegal instruction exception.
An implementation may have WFI always raise an illegal instruction
exception in less-privileged modes when TW=1, even if there are pending
globally-disabled interrupts when the instruction is executed.